requirements.txt
pytest
allure-pytest
appium-python-client
selenium
